ru.readmanga: moved helper functions into separate file
