+++ src/rust/ru.readmanga/src/helpers.rs
@@ -1,0 +1,35 @@
+use aidoku::{
+	error::Result,
+	prelude::*,
+	std::net::{HttpMethod, Request},
+	Manga, MangaPageResult,
+};
+use alloc::{string::String, vec::Vec};
+
+use crate::{
+	constants::{BASE_URL, SEARCH_OFFSET_STEP},
+	wrappers::WNode,
+};
+
+pub fn get_html(url: &str) -> Result<WNode> {
+	Request::new(url, HttpMethod::Get)
+		.html()
+		.map(WNode::from_node)
+}
+
+pub fn get_manga_url(id: &str) -> String {
+	format!("{}/{}", BASE_URL, id)
+}
+
+pub fn create_manga_page_result(mangas: Vec<Manga>) -> MangaPageResult {
+	let has_more = mangas.len() == SEARCH_OFFSET_STEP as usize;
+	MangaPageResult {
+		manga: mangas,
+		has_more,
+	}
+}
+
+pub fn get_chapter_url(manga_id: &str, chapter_id: &str) -> String {
+	// mtr is 18+ skip
+	format!("{BASE_URL}/{manga_id}/{chapter_id}?mtr=true")
+}
