+++ src/rust/ru.readmanga/src/lib.rs
@@ -3,6 +3,7 @@
 #![feature(iter_intersperse)]
 
 mod constants;
+mod helpers;
 mod parser;
 mod sorting;
 mod wrappers;
@@ -38,9 +39,9 @@ pub fn initialize() {
 #[get_manga_list]
 pub fn get_manga_list(filters: Vec<Filter>, page: i32) -> Result<MangaPageResult> {
 	let search_url = parser::get_filter_url(&filters, &Sorting::default(), page)?;
-	let html = parser::get_html(&search_url)?;
+	let html = helpers::get_html(&search_url)?;
 	let mangas = parser::parse_search_results(&html)?;
-	let result = parser::create_manga_page_result(mangas);
+	let result = helpers::create_manga_page_result(mangas);
 	debug!("get_manga_list: {result:?}");
 
 	Ok(result)
@@ -50,9 +51,9 @@ pub fn get_manga_list(filters: Vec<Filter>, page: i32) -> Result<MangaPageResult
 pub fn get_manga_listing(listing: Listing, page: i32) -> Result<MangaPageResult> {
 	let sorting = Sorting::from_listing(&listing);
 	let url = parser::get_filter_url(&vec![], &sorting, page)?;
-	let html = parser::get_html(&url)?;
+	let html = helpers::get_html(&url)?;
 	let mangas = parser::parse_search_results(&html)?;
-	let result = parser::create_manga_page_result(mangas);
+	let result = helpers::create_manga_page_result(mangas);
 	debug!("get_manga_listing: {result:?}");
 
 	Ok(result)
@@ -60,8 +61,8 @@ pub fn get_manga_listing(listing: Listing, page: i32) -> Result<MangaPageResult>
 
 #[get_manga_details]
 pub fn get_manga_details(manga_id: String) -> Result<Manga> {
-	let url = parser::get_manga_url(&manga_id);
-	let html = parser::get_html(&url)?;
+	let url = helpers::get_manga_url(&manga_id);
+	let html = helpers::get_html(&url)?;
 	let result = parser::parse_manga(&html, manga_id);
 	debug!("get_manga_details: {result:?}");
 
@@ -70,8 +71,8 @@ pub fn get_manga_details(manga_id: String) -> Result<Manga> {
 
 #[get_chapter_list]
 pub fn get_chapter_list(manga_id: String) -> Result<Vec<Chapter>> {
-	let url = parser::get_manga_url(&manga_id);
-	let html = parser::get_html(&url)?;
+	let url = helpers::get_manga_url(&manga_id);
+	let html = helpers::get_html(&url)?;
 	let result = parser::parse_chapters(&html, &manga_id);
 	debug!("get_chapter_list: {result:?}");
 
@@ -80,8 +81,8 @@ pub fn get_chapter_list(manga_id: String) -> Result<Vec<Chapter>> {
 
 #[get_page_list]
 pub fn get_page_list(manga_id: String, chapter_id: String) -> Result<Vec<Page>> {
-	let url = parser::get_chapter_url(&manga_id, &chapter_id);
-	let html = parser::get_html(&url)?;
+	let url = helpers::get_chapter_url(&manga_id, &chapter_id);
+	let html = helpers::get_html(&url)?;
 	let result = parser::get_page_list(&html);
 	debug!("get_page_list: {result:?}");
